requirements.txt
# Core
python-telegram-bot==20.6  # Modern async Telegram API
httpx[http2]>=0.24.0  # Faster HTTP client (alternative to requests)
uvicorn[standard]>=0.23.0  # ASGI server (for webhooks)
python-dotenv>=1.0.0  # Environment management

# Caching & Rate Limiting
redis>=4.5.5  # Redis client (async support)
aioredis>=2.0.0  # Async Redis (alternative)
tenacity>=8.2.2  # Retry logic

# Utilities
loguru>=0.7.0  # Structured logging
orjson>=3.9.0  # Fast JSON parsing
pydantic>=2.0.0  # Data validation
typing-extensions>=4.5.0  # Type hints support

# Dev & Testing (optional)
pytest-asyncio>=0.21.0  # Async tests
pytest-xdist>=3.3.0  # Parallel tests
ipython>=8.12.0  # REPL debugging
telegram>=0.0.1